python/basics/19 - managing keys/requirements.txt
blessings==1.7
certifi==2019.3.9
chardet==3.0.4
get==2019.4.13
idna==2.8
inquirer==2.5.1
post==2019.4.13
prompt-toolkit==2.0.9
public==2019.4.13
Pygments==2.4.0
python-dotenv==0.10.3
query-string==2019.4.13
readchar==2.0.1
regex==2019.4.14
request==2019.4.13
requests==2.22.0
six==1.12.0
urllib3==1.25.2
wcwidth==0.1.7
whaaaaat==0.5.2
windows-curses==1.1
